TbspRpgLib
Breaking up the service communication object, wip.
