+++ TbspRpgLib/InterServiceCommunication/RequestModels/Request.cs
@@ -1,0 +1,7 @@
+namespace TbspRpgLib.InterServiceCommunication.RequestModels
+{
+    public class Request
+    {
+        
+    }
+}
