+++ TbspRpgLib/LibStartup.cs
@@ -41,6 +41,7 @@ namespace TbspRpgLib {
             services.AddScoped<IAggregateTypeService, AggregateTypeService>();
 
             services.AddScoped<ITokenManager, TokenManager>();
+            services.AddScoped<IServiceManager, ServiceManager>();
             services.AddScoped<IServiceCommunication, ServiceCommunication>();
             services.AddScoped<IAdventureServiceLink, AdventureServiceLink>();
             services.AddScoped<IUserServiceLink, UserServiceLink>();
